Fixed warnings in rendergraph editor

+++ LambdaEngine/Include/Rendering/RenderGraphEditor.h
@@ -47,7 +47,7 @@ namespace LambdaEngine
 		void InternalRenderEditResourceView(RenderGraphResourceDesc* pResource, char* pNameBuffer, int32 nameBufferLength);
 
 		void RenderShaderView(float textWidth, float textHeight);
-		void RenderShaderTreeView(const LambdaDirectory& dir, float textWidth, float textHeight, int32& selectedIndex);
+		void RenderShaderTreeView(const LambdaDirectory& dir, float textWidth, float textHeight, int64& selectedIndex);
 
 		void RenderGraphView();
 		void RenderAddRenderStageView();
